Add download option to download all files from rm

+++ sync.py
@@ -7,6 +7,7 @@ parser = argparse.ArgumentParser(description='Synchronize reMarkable and Zotero 
 parser.add_argument('Zotero', help="Basefolder defined in ZotFile tablet settings.")
 parser.add_argument('reMarkable', help="Mirrored folder on reMarkable.")
 parser.add_argument('--delete', action='store_true', help="Delete all files on reMarkable that don't exists in the Zotero folder.")
+parser.add_argument('--download', action='store_true', help="Download all files from reMarkable, even if they are not in the Zotero folder. (Overwrites --delete)")
 parser.add_argument('--rmapi', default="rmapi", help="rmapi executable. This can be a path or the name of the rmapi executable that is in PATH. Default is rmapi")
 
 args = parser.parse_args()
@@ -46,9 +47,13 @@ def get_files():
     files_on_local = set([os.path.splitext(os.path.basename(f))[0] for f in os.listdir(ZOTERO_FOLDER)])
     return files_on_remarkable, files_on_local
 
-def process_files(delete=False):
+def process_files(delete=False, download=False):
     files_on_remarkable, files_on_local = get_files()
-    files_to_download = files_on_remarkable & files_on_local
+    files_to_download = files_on_remarkable
+
+    if not download:
+        files_to_download &= files_on_local
+
     files_to_upload = files_on_local - files_on_remarkable
     files_to_delete = files_on_remarkable - files_on_local
 
@@ -60,9 +65,9 @@ def process_files(delete=False):
         print(f"[Downloading] '{file}'")
         download_file(file)
 
-    if delete:
+    if delete and not download:
         for file in files_to_delete:
             print(f"[Deleting   ] '{file}'")
             delete_file(file)
 
-process_files(delete=args.delete)
+process_files(delete=args.delete, download=args.download)
